chore: more refactor
