+++ src/compiler/bounding-box.test.ts
@@ -1,5 +1,5 @@
 import type { GoslingSpec, Track } from '@gosling-lang/gosling-schema';
-import { DEFAULT_CIRCULAR_VIEW_PADDING, DEFAULT_VIEW_SPACING } from '../core/defaults';
+import { DEFAULT_CIRCULAR_VIEW_PADDING, DEFAULT_VIEW_SPACING } from './defaults';
 import { getBoundingBox, getRelativeTrackInfo } from './bounding-box';
 import { traverseToFixSpecDownstream } from '../compiler/spec-preprocess';
 import { getTheme } from '../core/utils/theme';
